+++ school_app/fl_components/README.md
@@ -1,2 +1,11 @@
 # PRUEBAGIT
-UNA JODIDA PRUEBA
+Este proyecto fue creado con el fin de minimizar la carencia de informacion para los alumnos de
+la unam y facultad de ingenieria creando un repositorio con reedireccion inmediata a los 
+principales citios web que un alumno deberia de conocer.
+
+AUTORES:
+Bryan Pineda Luna
+Steven Gonzalez Tapia
+Ricardo Hernandez Hernandez
+
+
