+++ school_app/fl_components/lib/main.dart
@@ -3,9 +3,6 @@
 import 'package:fl_components/screens/screens.dart';
 import 'package:flutter/material.dart';
 
-import 'package:fl_components/theme/app_theme.dart';
-import 'package:fl_components/router/app_routes.dart';
-
 void main() => runApp(const MyApp()); //*funcion principal "arranque de la app"
 
 //*esta es la clase principal
