--- school_app/fl_components/lib/models/menu_option.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart'; //show IconData, Widget
-
-class MenuOption {
-  final String route;
-  final Widget screen;
-
-  MenuOption({
-    required this.route,
-    required this.screen,
-  });
-}
