--- school_app/fl_components/lib/models/models.dart
@@ -1,1 +1,0 @@
-export 'package:fl_components/models/menu_option.dart';
