+++ school_app/fl_components/lib/screens/contenedores.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:fl_components/screens/screens.dart';
 import 'package:flutter/material.dart';
-import "package:fl_components/router/app_routes.dart";
 
 class contenedores extends StatelessWidget {
   double ancho = 0;
