+++ school_app/fl_components/lib/screens/contenedores2.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:fl_components/screens/screens.dart';
 import 'package:flutter/material.dart';
-import "package:fl_components/router/app_routes.dart";
 
 class contenedores2 extends StatelessWidget {
   double ancho = 0;
