+++ school_app/fl_components/lib/screens/screens.dart
@@ -1,7 +1,7 @@
 //? aqui estan todas las paqueterias de los archvos de screens
-export 'package:fl_components/screens/listview1_screen.dart';
-export 'package:fl_components/screens/listview2_screen%20.dart';
+export 'package:fl_components/screens/facultading.dart';
+
 export "package:fl_components/screens/contenedores2.dart";
-export 'package:fl_components/screens/card_screens.dart';
+export 'package:fl_components/screens/unam.dart';
 export 'package:fl_components/screens/home_screen.dart';
 export "package:fl_components/screens/contenedores.dart";
